Reduce severity of orphaned change subscription entry
If a wiki that is subscribed to some entities is closed or deleted or
becomes otherwise unavailable, then the message touched here is logged.
A script should be run to remove the entries of that wiki from the
subscription table, see T292512.

However, that does not constitute an error, because we can still
dispatch to all enabled wikis just fine. Therefore this is downgraded to
a warning instead. Even downgrading it to a notice might be defendable,
though I'm not sure notices are actually logged in wikimedia production,
i.e. kafka and such, when they are thrown in jobs.

Change-Id: I34f99c5f03f57d9e0b42218ad912c146b5cdd71a