Start 1.38 release notes with change dispatch migration
This documents how to manage the change dispatching part of the next
release without major disturbances (especially on wikis that never set
up proper change dispatching before, so they might have tons of
wb_changes rows that were never pruned).

The prose is expected to be refined/polished prior to the actual
release, but this documents the steps already.

The RELEASE-NOTES-1.37 are left as they are for now (not moved into
HISTORY yet), since they don’t look done yet.

Bug: T291162
Change-Id: Ibba7ddc5803accdb5fabd1515779c1428bfbcbc2