Don't filter by change Id when dispatching to client wikis
Filtering by change ID was only needed while the new system ran in
parallel with the legacy change dispatching and prune done by
maintenance scripts. Especially the fact that rows wb_changes were
only pruned after three days made it necessary to introduce some
mechanism to not dispatch those old changes again and again. This is no
longer needed.

This commit includes the first part of the patch that no longer requires
the change id in the job's methods. This change needs to be deployed to
all job runners, before the follow-up patch, that then includes calling
those methods without the change id, is deployed to the app servers

Change-Id: I95f560ffcab4ab7c507ef3615d927f65d7d1ffc6