Remove traces of dispatchingLockManager repo option
This removes the option’s default (it was already no longer read
anywhere) and its documentation in options.md. It’s still mentioned in
change-propagation.md – that document needs a complete overhaul, I don’t
think it makes sense to remove just the mention of the setting there at
the moment.

Bug: T292604
Change-Id: I36631e8fbfa226c810b2d0ad19477c2157849a2d