Document requirement on item and property source
The WikibaseClient.ExternalUserNames wiring uses the item and property
source’s database name as a site ID, and if that site is not known to
MediaWiki’s SiteLookup, then the RecentChangesFactory will try to create
RecentChange objects for unprefixed user names, which MediaWiki won’t
allow. Document that the database name must therefore be a site ID for
the repo wiki, in the regular docs and also in the release notes.

Bug: T293246
Change-Id: I7bd7be2bd9b30f71160dd951ab83854ee86a6194