Catch CannotCreateActorException in InjectRCRecords
In our investigation so far, this happens when the ExternalUserNames
service is null (because the item and property source has a string
database name that’s not known as a site global ID, or because that site
has no interwiki prefixes), so we try to create a recentchanges entry
with rc_user being 0 but the rc_user_text being an unprefixed, usable
name, which isn’t allowed (the ActorStore refuses to acquire such an
actor – after all, a user with that name might actually register later).
Catch this error and log it, but without crashing the whole job (since
it’s unlikely to work when retrying).

Bug: T293246
Change-Id: I3c0734235a169c71d9a43ac371702238add8968d