Minor update to the package-lock.json format
Some of the more recent npm7 releases must have introduced a slightly
different way of storing some information in the package-lock.json. This
introduces noise to commits that actually change the version of a
package, so it is done here in a separate commit.

This was generated with Node.js v12.21.0 and npm 7.24.2

Change-Id: Ie4fd134ee951b5331e7f3a9e314c527fc4dc84a4