bridge: Switch to Vue3 compat
This is the next step for migrating data-bridge to Vue 3 for T294465.
The changes done in this patch are necessary so that the app builds
correctly, passes all tests and storybook works.
This is now using the vue 3 compatibility build. Future commits will
have to address all the deprecation messages.

Since we also have to upgrade vuex-smart-modules, we also have to switch
to Vuex 4 and ship it ourselves again until it is provided by core.

Change-Id: I1822878050997eedbd2419fd5920c96d1bd65172