Stop calling WikiPage::factory() in WikiPageEntityStore
Instead, inject a WikiPageFactory. SqlStore still gets it from the
global service container (instead of having it injected as well), but
it’s a start.

While we’re at it, clean up some of the tests to use the test’s service
container instead of relying on the global one (they should be the same,
but this is more elegant).

Bug: T297706
Change-Id: Iad1cf709b4cb30513caf97653ff3de11b30df7f7