bridge: Get rid of in-language directive
Using the directive TermLabel in Vue3 seems to cause unsolvable problems
in e2e testing.

However, this causes problems in the e2e test. This plugin modifies the
global Vue instance. This global vue instance is shared and thus
replacing the language related mocks on the global window instance
doesn't have the desired effect. This problem will solve itself with
switching to Vue3 as then there is no longer a global Vue instance to be
mutated, but rather the specific App instance is mutated and that is
recreated for each test.

Change-Id: Iee63dd1e9c6c840d542a586f8ffcbaa826f55780