Rename variables of type ParserOutput to disambiguate from OutputPage
When refactoring or otherwise maintaining ParserOutput (for example,
T287216) it is often hard to distinguish instances of OutputPage from
instanceof of ParserOutput, especially since they are often both
stored in variables named $output or $out and contain many
identically-named methods.

As an aid to future maintainers, code search, and human readers, try to
consistently use $parserOutput to refer to variables storing ParserOutput
instances to avoid confusion.

Change-Id: I9fdeb505b5190e6f82536024b1ef3c358f8078e3