Support idGenerator=auto and use it in CI
The default / original ID generator can cause deadlocks in CI when tests
are run in parallel. Introduce a new setting value, “auto”, to choose
between the two available implementations depending on the database
type, and use that value in CI. (The non-CI default remains “original”
for now.)

Note that we specifically use a lazy database connection to determine
the database type, and DBConnRef overrides getType() to avoid triggering
a database connection, so the service wiring should not cause any
connection to be made yet.

Bug: T298682
Change-Id: I456a23cd408680bc416ed506b5130f4224214511
Depends-On: I0e101e4134123cb6157dd175c139014f5578ce55