Simplify LanguageFallbackChainFactory implementation
Previously, buildFromLanguage() could be called with an arbitrary
bitmask of FALLBACK_* flags, and in practice the following combinations
were possible (after we removed $mode from the public interface):

* FALLBACK_ALL, from newFromLanguage(Code)
* FALLBACK_SELF | FALLBACK_VARIANTS, from recursive call
* FALLBACK_SELF, from buildFromBabel()
* FALLBACK_VARIANTS, from buildFromBabel()
* FALLBACK_OTHERS | FALLBACK_VARIANTS, from buildFromBabel()

One important observation here is that, while buildFromBabel() calls
buildFromLanguage() with “exclusive” modes (“self and nothing else”,
“variants but not self”, “others with variants but not self”), it only
does this to get the codes into the chain in a certain order. In the
end, all the different modes are supposed to be included in the chain –
the Babel code just wants “self” before “variants”, and “variants”
before “others”, for several language codes (e.g. the “self”s of several
languages before the “variants” of any of them). This means that we can
simplify this to three separate methods:

* addLanguageToChain() = FALLBACK_SELF
* addLanguageAndVariantsToChain() = FALLBACK_SELF | FALLBACK_VARIANTS
* addLanguageAndVariantsAndFallbacksToChain() = FALLBACK_SELF |
  FALLBACK_VARIANTS | FALLBACK_OTHERS

When buildFromBabel() calls addLanguageAndVariantsToChain(), the “self”
part will actually be a no-op (because addLanguageToChain() will already
have been called for this language code at this point), but we can just
ignore that, since it greatly simplifies the code to have only three
methods, where the second and third method are implemented using the
first and second method – instead of having a dedicated method or mode
that really only adds “variants” without “self”, which was possible (but
never needed) in the old code.

The buildFromLanguage() method is now just a shortcut for calling
addLanguageAndVariantsAndFallbacksToChain() in a more convenient way (it
returns the chain instead of taking an “output” parameter). A follow-up
change will slightly expand it again.

Also, to reduce code duplication in buildFromBabel(), let’s call
LanguageWithConversion::validateLanguageCode() only once, so that we
don’t have to copy the whole try-catch block all the time.

Bug: T299663
Change-Id: I85648a8f082be6014cb0458175fb358179ed36ee