Don’t log any errors when SSR URL is not configured
This kind of extends Ia67f0deb03, in that the ssrServerUrl is not only
allowed to be null (and defaults to that), but we don’t even log any
kind of error in that case any more.

The TermboxRemoteRenderer callers (TermboxView::renderTermbox() and
ExternallyRenderedEntityViewPlaceholderExpander::rerenderTermbox())
catch the general exception type TermboxRenderingException without
logging anything, so by disabling the logging inside
TermboxRemoteRenderer for TermboxNoRemoteRendererException specifically,
we effectively disable logging for these errors.

Change-Id: I55c6133cd406d28277091d17910100413377d514