SpecialUnconnectedPages: use less strings for SQL
The namespace(s) are known to be ints, so I don’t think this was an SQL
injection issue or anything, but still, it’s nicer to let MediaWiki do
the SQL formatting when it’s possible, and it’s certainly possible here.

The phpdoc return type of buildConditionals() is no longer accurate (we
return either an int[] or an int[][]), so let’s remove that and just add
a static “array” return type.

In the tests, let’s use assertSame() to guarantee the correct element
type in the arrays.

Change-Id: I6a99a44a923faa547b2d2c5818ba3f34239e60a2