Store `noexternallanglinks` in extension data, not page property
We are moving towards an append-only model for the ParserOutput, and
updating the page properties multiple times is inconsistent with that
model (T300979/T300986).  Use the new `ParserOutput::appendExtensionData()`
method to safely accumulate the values from the various parser function
invocations.  This could be transferred to a page property later, but
it appears that it is not actually necessary to store the
`noexternallanglinks` data persistently as a page property; storing it
in the ParserOutput cache ought to be sufficient.

For backward compatibility, for now also check the page property in the
ParserOutput, in case we get stale contents from the parser cache.

Bug: T300986
Depends-On: I2dea4ba71ea506428854a9983c1abd906b2efd5f
Change-Id: I7df327a7f343d10b4a4f2ce48c69777b20af91b3