Enable data value normalization by default
And remove the temporary tmpNormalizeDataValues setting. Also, document
the feature in the 1.42 release notes; in principle, it could be
backported (in which case we should adjust the release notes entry on
the master branch later), but given how long it took to get this merged
on master, I’m not holding my breath for backports.

In EditEntityTest, we need to change the way we set the settings so that
the service container is reset (happens in setService()), which was
previously “masked” by the similar code in WikibaseApiTestCase. This
could probably be improved further (e.g. another test function in the
same class, testGivenReadOnlyType_errorIsShownAndNoEditHappened(), still
uses WikibaseRepo::getSettings()->setSetting()), but not right now.

Bug: T292181
Change-Id: I3d60fd16bc839e08623347ea20dc337d60873c99