Change a few Html::rawElement() to Html::element()
These are harmless (the NonExistingEntityIdHtmlBrokenLinkFormatter is
mostly harmless, but it would be quite a feat to get a malicious entity
ID to that point), but rawElement() makes me nervous and element() works
just as well here. It’s possible that rawElement() was meant as a
micro-optimization here (we know the string is safe, so it’s not worth
escaping it), but I don’t think that’s worth the miniscule performance
improvement.

Change-Id: I3fbe45624ce5dc20f79aeb92d81ffc794d2c12e0