tests: Don't rely on 'en' falling back to 'qqx' when set as site language
The expected messages are coming from
ApiErrorReporter::getPlainErrorMessageFromStatus(), which hardcodes
the message language to 'en'. However, because the site language was
set to 'qqx', and the message does not exist in 'en', the result was
in 'qqx'.

This doesn't work with I45f58034117e9041fe5403dbfcf83ba25f20518d,
which doesn't expect qqx messages to be used as fallbacks. I don't
think setting site language to 'qqx' should be supported.

Change-Id: I0a0681b332abb623fc6871ab78da9e64a79fad84