Add 'display' to wbsearchentities response
This indicates the language of the returned label and description, which
wasn’t previously available (it’s not necessarily as the search
language). In Federated Properties, we use the new part of the remote
response if it’s present, but fall back to the old 'label' and
'description' with the search language as the language code if we have
to (we don’t want to break support for older Wikibases) – using the
search language code might not be correct, but it’s the best we can do
and matches what the code did before (though I think the old code’s term
language was never used anywhere).

Bug: T104344
Change-Id: I40cd43cb126ce1abc7a4e638399e671292529ce1
Co-Authored-By: Michael Große <michael.grosse@wikimedia.de>