Use default scope in example settings
Wikibase.examples.php has been using the `global`keyword to access
the $wgWBRepoSettings configuration variable. However, with
Ifec91445e11, configuration variables are not global. They are copyied
into global variables later for backwards compatibility, but they are
not global variables while LocalSettings.php is being executed.

This patch changes Wikibase.examples.php to access $wgWBRepoSettings
in the file's scope, so the change can be picked up by
LocalSettingsLoader if that is being used. When LocalSettings.php is
included in global scope (the current default), nothing should change.

(A similar change was previously proposed in Ifec91445e11)

Needed-By: I0f20024803336064ecb07b51fa56581af7b67a85
Change-Id: I7a59bf8e7bbae5738f72eb7c5424e5b0dc403e9a