Use new display part of wbsearchentity response in entity selector
This lets us correctly language-tag the label and description if
language fallback happened. Note that the part in parentheses is still
not tagged correctly, since it uses the .aliases instead of the .match;
it also hard-codes an ASCII space and ASCII parentheses, which is not
correct in all languages. Fixing that is left for later.

The code still falls back to the old label/description fields – not so
much for compatibility with API responses from older Wikibase versions,
but rather because these “entity stubs” can also be provided by other
JavaScript code via the wikibase.entityselector.search hook, including
but not limited to the WikibaseQualityConstraints gadget that suggests
qualifier properties or item values based on constraints. Requiring
these “entity stubs” to contain a display section instead of a label and
description would be a breaking change (JS hooks are stable!).

Bug: T104344
Change-Id: Ifa79c159a2d185fd2e9ce1b254d2a8a56d23fbbb