statementlistview should not overlap the property
Make wikibase-statementlistview no longer overlap
wikibase-statementgroupview-property, by adding a
margin-left of 15em to it. This seems more logical
and cleaner than the previous approach (which is
adding that 15em margin to various child elements
individually).

I tested this in Firefox and Chromium and couldn't
spot any visual difference (except for the bug this
fixes, of course).
I also superficially tested the the mobile view
in the responsive design mode of the aforementioned
browsers.

Bug: T285694
Change-Id: I8f004240cf810fd06565ed478d99b77aed19f09e