Return Status for editing non-local entity instead of throwing
In change Iece8393833 (commit 63ba4d2fa3), I mistakenly made
MediawikiEditEntity::attemptSave() throw for non-local entities, based
on the existing code that throws a ReadOnlyError in similar conditions.
However, while a ReadOnlyError has special handling to be shown as an
API error, this isn’t the case for an InvalidArgumentException, so that
just resulted in an internal error. Instead, let’s return a bad Status
object with a message indicating the error.

To keep PHPCS from complaining about the cyclomatic complexity of the
method (it’s at the limit), merge two equivalent conditions into one
further down in the method.

Bug: T242959
Change-Id: I626a4c97f9d52d1eff01bd569437cba5f6817649