Add missing ORDER BY to UnexpectedUnconnectedPagePrimer
Since we’re going through the table in page ID order (page_id >
$this->position, page_id <= $lastPosition), the results must be ordered
by page ID; if we don’t require this explicitly, then we’re implicitly
relying on the database choosing to run the query in such a way that the
results are ordered by page ID, but that doesn’t always happen. If the
database runs the query in a different way, the last page ID returned
(becoming the $this->position) might happen to be much higher than
expected, and we would skip large parts of the table.

Also, change the update key in LoadExtensionSchemaUpdatesHookHandler, so
that the update runs again in update.php (the buggy version already ran
on the Beta cluster).

Change-Id: Ia0d2f9e5289debba70263af199972e43e83deee9