Fix outline for selected statement group
Since change I8f004240cf (commit 996a99c2bb), we need this outline to
apply to the whole .wikibase-statementgrouplistview, not only the
.wikibase-statementlistview inside it (which no longer spans the full
width of the statement group). However, for the right edge of the
statement to be visible, we also need the statement list view to be a
bit less wide – the width of the outline, to be exact. Extract that
width into a variable to make the relationship clear; to limit the scope
of the variable, extract the common .wikibase-statementgrouplistview
selector into a nested block.

Bug: T285694
Change-Id: I17e92bebd0ebaa082d250a1f7946c30c88005c38