REST: Disable auth api tests
We discovered that in case the test runs with the OAuth extension
enabled, and with apache as the webserver, it fails due to the
Authorization header getting stripped from the request. See T305143.
This was not discovered because the precondition for the OAuth extension
to be installed was not met in Wikibase CI.

Change-Id: Ib1b134d13ee23436e40557dcf2e8cffa76937eea