Sort Special:UnconnectedPages ascending in MIGRATION_NEW
Ideally, we’d like to sort by ascending namespace, but descending page
ID, but that’s not efficient to do (unless we invert the page prop to
contain the negative page ID). Since the __EXPECTED_UNCONNECTED_PAGE__
magic word is now available, and the expected ideal state of this
special page is for it to be completely empty, we think it’s also
acceptable to sort by ascending page ID instead, so that the article
namespace still comes first. This could be reconsidered if necessary.

Change-Id: I362ff336eebcf1939caa202eb5fe196df3c8a8f6