Don’t use session-consistent connections in UnexpectedUnconnectedPagePrimer
We don’t want getUnexpectedUnconnectedBatch() to start connecting to the
primary after the first insertUnexpectedUnconnectedBatch(). For one
thing, we already wait for replication after each batch; but even if we
didn’t, we generally accept that there might be a race between the page
being selected as unexpected unconnected and the page prop being
written, whether this is due to a concurrent write or due to replication
lag (that’s also why we use INSERT IGNORE). The query to select a batch
of unexpected unconnected pages can be costly, and we want a replica to
bear that burden, not the primary.

Change-Id: Icaf01340140225991123d2119f4d6b1a06b4e8bf
(cherry picked from commit 73434c0cb8f9c56393fad000f3cd3e48c22d1d65)