Check content, not title, for redirect status
At the point where we check for an unexpected unconnected page, the
redirect status might not be up to date in the database, it seems. Let’s
pass through the $content (which the hook handler already receives) and
ask it whether it’s a redirect, rather than having the $title look it up
in the database.

Change-Id: I479622cd421b40ce377671223f43c5596a4bd837