Fix EntityHandlerTestCase::testPageProperties()
Asserting that the parser output has text when we are asking the content
renderer to not generate HTML (final argument, false) doesn’t make much
sense, and fails with MediaWiki core change I49cdf21411.

To avoid a warning about the “empty” version of this test not making any
assertions, expand the test a bit so it can assert the absence of page
props, and make the “empty” test case assert that the wb-claims page
prop isn’t there.

Bug: T306591
Change-Id: I9390a3fcc9d68040475275f6e1303122a9ea136f
Needed-By: I49cdf21411c6b02ac9a221a13393bebe17c7871e