Don’t assert text of parser output without HTML
The fact that $parserOutput->getText() on a ParserOutput without HTML
returns the empty string, instead of throwing “This ParserOutput
contains no text!”, is arguably a bug, and will be fixed in MediaWiki
core. Instead, first try asserting that the ParserOutput has no text at
all; once the MediaWiki core change is merged, we can drop the second
part of the assertion.

Bug: T306591
Change-Id: Ie5fb5da50eb2e32bc9a5e2eadef6c957f635f3c9
Needed-By: I49cdf21411c6b02ac9a221a13393bebe17c7871e