Restore additional page props tests
For items, this method was missed when the page props test moved from
EntityContentTestCase to EntityHandlerTestCase in change I1037978952
(commit e8e9be4182). Move it over so it can override the parent method
again, adjusting the implementation a bit to fit the new environment
($this->newEntityContent() replaces $this->newBlank( $this->getDummyId() )).

For properties, this method didn’t exist before, but there was a comment
in the item version saying that this should exist once we support
statements on properties. We’ve supported that for many years now
(T51554, late 2014), so let’s add a test case.

Change-Id: Ie75f972256a946d3f6de1fd8bbbf85b234a11367