Use *Ref() connections instead of releaseConnection()
The non-Ref() ConnectionManager methods are deprecated, and so is
releaseConnection(): it’s better to use the Ref() version, which
automatically releases the connection when the variable goes out of
scope.

Three calls to releaseConnection() are still left, corresponding to
get*Connection() calls with a $flags argument, which isn’t supported by
the Ref() methods yet. Those will have to be fixed later.

Change-Id: Ia5a0fdc2f1859c4fc313c4ea36350ea510ba72de