Simplify InfoActionHookHandler::formatEntityUsage()
Years ago, this code was copied to EditActionHookHandler in change
I1f1fad7855 (commit 6d024aeb44), and some improvements were made in the
process; most significantly, the EntityIdParser became unused (though I
only removed it from the class just now in change Ic5d988fb36). However,
these improvements were never ported back to InfoActionHookHandler. This
change makes up for that.

The specific changes are: rename $entities to $entityIds (it always
contained IDs, not entity documents), and use it everywhere so that we
never need to re-parse entity IDs; rename $usage to $usages (it was
always an array), and clarify its phpdoc; add a static return type; swap
the two arguments, purely for consistency with EditActionHookHandler;
and remove the now-unused EntityIdParser.

With this change, the beginning of formatEntityUsage() is identical in
EditActionHookHandler and InfoActionHookHandler. (The second half,
starting after the $text assignment, is different, because the two hook
handlers show the usage information in different ways.)

Change-Id: Ic433fd3369f22f58a8d6ca0eb44282b5bdd20334