Default unexpectedUnconnectedPage migration to MIGRATION_WRITE_BOTH
This will cause the new page prop to be written in a default Wikibase
install (including, via update.php, for existing pages), and prepares
for a later change of the default to MIGRATION_NEW, and then to use the
new page prop unconditionally.

While we’re at it, also correct the comment – the way that
SpecialUnconnectedPages is actually implemented, only MIGRATION_NEW will
cause the new page prop to be read, MIGRATION_WRITE_NEW is not enough.
(Arguably this is a bug, but not one that’s worth fixing now, IMHO.)

Change-Id: I3235fd2d24d0e545f15b1d536a273c98377e1474