Fix wb.getUserLanguages() if user language not in babel
If the wgUserLanguage is not in the wbUserSpecifiedLanguages, then
indexOf will return -1 and splice will interpret that as an index from
the end of the array and remove the last element. Fix that to only call
splice if the user language is actually in the languages.

Also add a test for the intended behavior behind this splice() call,
which is to move the user interface language to the front of the list,
and which (I think) wasn’t really tested so far.

Bug: T307808
Change-Id: I59573c63e6b469f0467be4a3ccd14c06ce9ad6aa