REST: Serialize in presenter for GetItem
This moves the serialization from the GetItem use case into the
GetItemJsonPresenter. This split required a new mechanism for the use
case to transfer partial item data in its response object, which is what
the ItemData class was created for.

For now the ItemData object gets created by the use case itself given
the fields list from the request object and the Item from the
ItemRevisionRetriever. This isn't ideal, but will soon be refactored out
as part of the work on T307915 which will introduce a dedicated
ItemDataRetriever service.

The serialization in the presenter now happens via the
ItemDataSerializer which combines the custom serialization for terms
with existing serializers for statements and sitelinks.

Bug: T308437
Change-Id: Ia63dfd1898489fbabfa6d82ea6ed2083bf2ff5f2